+++ src/core/application_builder.py
@@ -5,6 +5,7 @@ import tkinter as tk
 from typing import TYPE_CHECKING
 
 from src.gui.theme_manager import ThemeManager
+from src.services.fixed_phrases_manager import FixedPhrasesManager
 from src.utils.error_handler import log_and_show_error
 from src.utils.i18n import Translator
 
@@ -14,13 +15,12 @@ from .config.settings_manager import SettingsManager
 from .dependency_checker import DependencyChecker
 from .event_dispatcher import EventDispatcher
 from .exceptions import ConfigError
-from .fixed_phrases_manager import FixedPhrasesManager
 from .plugin_manager import PluginManager
 
 if TYPE_CHECKING:
     from src.core.app_main import MainApplication
-    from src.core.history_service import HistoryService
     from src.db.database_manager import DatabaseManager
+    from src.services.history_service import HistoryService
 
 logger = logging.getLogger(__name__)
 
@@ -111,7 +111,7 @@ class ApplicationBuilder:
         if not self.event_dispatcher or not self.db_manager:
             raise ConfigError("イベントディスパッチャまたはデータベースマネージャーが初期化されていません")
         try:
-            from src.core.history_service import HistoryService
+            from src.services.history_service import HistoryService
             # デフォルト設定から履歴数上限を取得（後でSETTINGS_CHANGEDでも同期されます）
             limit = 50
             if self.settings_manager:
