+++ src/services/fixed_phrases_manager.py
