+++ src/services/notification_manager.py
