+++ tests/conftest.py
@@ -5,8 +5,8 @@ from collections.abc import Generator
 import pytest
 
 from src.core.event_dispatcher import EventDispatcher
-from src.core.history_service import HistoryService
 from src.db.database_manager import DatabaseManager
+from src.services.history_service import HistoryService
 
 
 @pytest.fixture
